RE #27037 Move observer pattern to common location
The observer pattern is not Muon specific and could be used elsewhere. Moving it to somewhere not related to the muon GUI reduces the chance of the code being duplicated.
