+++ qt/python/mantidqt/utils/observer_pattern.py
