+++ scripts/test/Muon/home_grouping_widget_test.py
@@ -16,7 +16,7 @@ from Muon.GUI.Common.home_grouping_widget.home_grouping_widget_model import Home
 from Muon.GUI.Common.home_grouping_widget.home_grouping_widget_presenter import HomeGroupingWidgetPresenter
 from Muon.GUI.Common.home_grouping_widget.home_grouping_widget_view import HomeGroupingWidgetView
 from Muon.GUI.Common.muon_pair import MuonPair
-from Muon.GUI.Common.observer_pattern import Observer
+from mantidqt.utils.observer_pattern import Observer
 from Muon.GUI.Common.test_helpers.context_setup import setup_context_for_tests
