+++ scripts/test/Muon/muon_gui_context_test.py
@@ -8,7 +8,7 @@ import sys
 import unittest
 
 from Muon.GUI.Common.contexts.muon_gui_context import MuonGuiContext
-from Muon.GUI.Common.observer_pattern import Observer
+from mantidqt.utils.observer_pattern import Observer
 from mantidqt.utils.qt.testing import start_qapplication
 
 if sys.version_info.major < 2:
